==> facilities/carf/01-carf-scrape.ipynb <==
# CARF accredited facilities

import pandas as pd

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.support.ui import Select

from webdriver_manager.chrome import ChromeDriverManager

import requests
from bs4 import BeautifulSoup

import re

driver = webdriver.Chrome(ChromeDriverManager().install())

driver.get("http://www.carf.org/advancedProviderSearch.aspx")

## All CARF accredited facilities that offer ED treatment

### Residential (BH)

# Select program: residential treatment (BH)
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlProgram"]/option[129]').click()

# Select age group/special population: eating disorders
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlPopulation"]/option[8]').click()

# Press search ctl00_body_btnAdvSearch
driver.find_element(By.ID, 'ctl00_body_btnAdvSearch').click()

# Pull into beautiful soup
doc = BeautifulSoup(driver.page_source)

# Create empty dataframe
data = []

for each_row in doc.find_all('tr')[10:20]:
    facility = {}
# name
    facility['name'] = each_row.find('a').text
# link
    facility['link'] = 'http://www.carf.org/' + each_row.find('a')['href']
# city
    facility['city'] = each_row.find_all('td')[1].text
# state
    facility['state'] = each_row.find_all('td')[2].text
# skip country column
# affiliated/parent company
    facility['company'] = each_row.find_all('td')[4].text
# type column, which will change for each dropdown
    facility['type'] = 'BH'
    
# click on link to get more info
    driver.get('http://www.carf.org/' + each_row.find('a')['href'])
    doc = BeautifulSoup(driver.page_source)
    try:
        facility['address'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[0].previousSibling
        facility['address_pt2'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[1].previousSibling
        facility['phone'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[3].previousSibling
        facility['website'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[4].previousSibling

    except:
        facility['address'] = 'NA'
        facility['address_pt2'] = 'NA'
        facility['phone'] = 'NA'
        facility['website'] = 'NA'
    
    data.append(facility)
    
len(data)

# go back to search page
driver.get("http://www.carf.org/advancedProviderSearch.aspx")

# Select program: residential treatment (BH)
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlProgram"]/option[129]').click()

# Select age group/special population: eating disorders
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlPopulation"]/option[8]').click()

# Press search ctl00_body_btnAdvSearch
driver.find_element(By.ID, 'ctl00_body_btnAdvSearch').click()

# click to page 2
driver.find_element(By.XPATH, '//*[@id="ctl00_body_gvProviderAdvSearch_ctl13_btnPagingNum2"]').click()

# pull into BS
doc = BeautifulSoup(driver.page_source)

# add to data

for each_row in doc.find_all('tr')[10:20]:
    facility = {}
# name
    facility['name'] = each_row.find('a').text
# link
    facility['link'] = 'http://www.carf.org/' + each_row.find('a')['href']
# city
    facility['city'] = each_row.find_all('td')[1].text
# state
    facility['state'] = each_row.find_all('td')[2].text
# skip country column
# affiliated/parent company
    facility['company'] = each_row.find_all('td')[4].text
# type column, which will change for each dropdown
    facility['type'] = 'BH'
    
# click on link to get more info
    driver.get('http://www.carf.org/' + each_row.find('a')['href'])
    doc = BeautifulSoup(driver.page_source)
    try:
        facility['address'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[0].previousSibling
        facility['address_pt2'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[1].previousSibling
        facility['phone'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[3].previousSibling
        facility['website'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[4].previousSibling

    except:
        facility['address'] = 'NA'
        facility['address_pt2'] = 'NA'
        facility['phone'] = 'NA'
        facility['website'] = 'NA'
        
    data.append(facility)
    
len(data)

# go back to search page
driver.get("http://www.carf.org/advancedProviderSearch.aspx")

# Select program: residential treatment (BH)
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlProgram"]/option[129]').click()

# Select age group/special population: eating disorders
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlPopulation"]/option[8]').click()

# Press search ctl00_body_btnAdvSearch
driver.find_element(By.ID, 'ctl00_body_btnAdvSearch').click()

# click to page 3
driver.find_element(By.XPATH, '//*[@id="ctl00_body_gvProviderAdvSearch_ctl13_btnPagingNum3"]').click()

# pull into BS
doc = BeautifulSoup(driver.page_source)

# add to data

for each_row in doc.find_all('tr')[10:12]:
    facility = {}
# name
    facility['name'] = each_row.find('a').text
# link
    facility['link'] = 'http://www.carf.org/' + each_row.find('a')['href']
# city
    facility['city'] = each_row.find_all('td')[1].text
# state
    facility['state'] = each_row.find_all('td')[2].text
# skip country column
# affiliated/parent company
    facility['company'] = each_row.find_all('td')[4].text
# type column, which will change for each dropdown
    facility['type'] = 'BH'

    # click on link to get more info
    driver.get('http://www.carf.org/' + each_row.find('a')['href'])
    doc = BeautifulSoup(driver.page_source)
    try:
        facility['address'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[0].previousSibling
        facility['address_pt2'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[1].previousSibling
        facility['phone'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[3].previousSibling
        facility['website'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[4].previousSibling

    except:
        facility['address'] = 'NA'
        facility['address_pt2'] = 'NA'
        facility['phone'] = 'NA'
        facility['website'] = 'NA'
        
    data.append(facility)
    
len(data)

### BH, children and adolescents

# go back to search page
driver.get("http://www.carf.org/advancedProviderSearch.aspx")

# Select program: residential treatment (BH)
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlProgram"]/option[129]').click()

# Select age group/special population: eating disorders, children and adolescents
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlPopulation"]/option[9]').click()

# Press search ctl00_body_btnAdvSearch
driver.find_element(By.ID, 'ctl00_body_btnAdvSearch').click()

# Pull into beautiful soup
doc = BeautifulSoup(driver.page_source)

for each_row in doc.find_all('tr')[10:20]:
    facility = {}
# name
    facility['name'] = each_row.find('a').text
# link
    facility['link'] = 'http://www.carf.org/' + each_row.find('a')['href']
# city
    facility['city'] = each_row.find_all('td')[1].text
# state
    facility['state'] = each_row.find_all('td')[2].text
# skip country column
# affiliated/parent company
    facility['company'] = each_row.find_all('td')[4].text
# type column, which will change for each dropdown
    facility['type'] = 'BH, children and adolescents'
    
# click on link to get more info
    driver.get('http://www.carf.org/' + each_row.find('a')['href'])
    doc = BeautifulSoup(driver.page_source)
    try:
        facility['address'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[0].previousSibling
        facility['address_pt2'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[1].previousSibling
        facility['phone'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[3].previousSibling
        facility['website'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[4].previousSibling

    except:
        facility['address'] = 'NA'
        facility['address_pt2'] = 'NA'
        facility['phone'] = 'NA'
        facility['website'] = 'NA'
        
    data.append(facility)
    
len(data)

# go back to search page
driver.get("http://www.carf.org/advancedProviderSearch.aspx")

# Select program: residential treatment (BH)
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlProgram"]/option[129]').click()

# Select age group/special population: eating disorders, children and adolescents
driver.find_element(By.XPATH, '//*[@id="ctl00_body_ddlPopulation"]/option[9]').click()

# Press search ctl00_body_btnAdvSearch
driver.find_element(By.ID, 'ctl00_body_btnAdvSearch').click()

# Pull into beautiful soup
doc = BeautifulSoup(driver.page_source)
    
# click to page 2
driver.find_element(By.XPATH, '//*[@id="ctl00_body_gvProviderAdvSearch_ctl13_btnPagingNum2"]').click()

# pull into BS
doc = BeautifulSoup(driver.page_source)

# add to data

for each_row in doc.find_all('tr')[10:15]:
    facility = {}
# name
    facility['name'] = each_row.find('a').text
# link
    facility['link'] = 'http://www.carf.org/' + each_row.find('a')['href']
# city
    facility['city'] = each_row.find_all('td')[1].text
# state
    facility['state'] = each_row.find_all('td')[2].text
# skip country column
# affiliated/parent company
    facility['company'] = each_row.find_all('td')[4].text
# type column, which will change for each dropdown
    facility['type'] = 'BH, children and adolescents'
    
# click on link to get more info
    driver.get('http://www.carf.org/' + each_row.find('a')['href'])
    doc = BeautifulSoup(driver.page_source)
    try:
        facility['address'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[0].previousSibling
        facility['address_pt2'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[1].previousSibling
        facility['phone'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[3].previousSibling
        facility['website'] = doc.find(id='ctl00_body_pnlProviderInformation').find_all('br')[4].previousSibling

    except:
        facility['address'] = 'NA'
        facility['address_pt2'] = 'NA'
        facility['phone'] = 'NA'
        facility['website'] = 'NA'
        
    data.append(facility)
    
len(data)

### Save to dataframe

# save into dataframe
df = pd.DataFrame(data)

# replace line breaks
df = df.replace('\n', '', regex=True)

# strip leading and trailing white space
for column in df.columns:
    df[column] = df[column].str.strip()
    
# remove phone and website text
df['phone'] = df.phone.str.replace('Phone: ', '', regex=False)
df['website'] = df.phone.str.replace('Website: ', '', regex=False)

df.head()

df.to_csv('carf-accredited-raw.csv', index=False)

## Services offered

data_programs = []
for each_entry in data:
    program = {}
    try:
        driver.get(each_entry['link'])
        doc = BeautifulSoup(driver.page_source)
        for each_row in doc.find_all('tbody')[0].find_all('tr')[1:]:
            program = {}
            program['name'] = each_entry['name']
            program['link'] = each_entry['link']            
            program['program_offered'] = each_row.find_all('td')[0].text + each_row.find_all('td')[1].text + each_row.find_all('td')[2].text
            data_programs.append(program)
    except:
        pass
    
len(data_programs)

### Save to dataframe

df_programs = pd.DataFrame(data_programs)

df_programs.program_offered.unique()

# save into dataframe
df_programs = pd.DataFrame(data_programs)

# replace line breaks
df_programs = df_programs.replace('\n', '', regex=True)

# strip leading and trailing white space
for column in df_programs.columns:
    df_programs[column] = df_programs[column].str.strip()
    
df_programs["program_offered"] = df_programs.program_offered.str.replace('\s+', ' ', regex=True)

df_programs.head()

df_programs.to_csv('programs-only-raw.csv', index=False)

# Will manually split out in Excel
