==> facilities/carf/02-merge.ipynb <==
import pandas as pd

df = pd.read_csv('carf-accredited-raw.csv')
df_programs = pd.read_csv('programs-only-clean.csv')

print(len(df))
print(len(df_programs))

# sort names a-z, reset the index alpabetically, drop old index
df = df.sort_values('name', ascending=True).reset_index().drop(['index'], axis=1)
df_programs = df_programs.sort_values('name', ascending=True).reset_index().drop(['index'], axis=1)

# rename name and link column names in df_programs to avoid duplicate col names in merge
df_programs = df_programs.rename(columns={"name": "name_programs", "link": "link_programs"})

# merge
df_merge = pd.concat([df, df_programs], axis=1)
len(df_merge)

# drop duplicate name and link columns - COMMENT OUT TO LEAVE IN THE DUPLICATE NAME & LINK COLUMNS TO DOUBLE CHECK THE CONCAT
df_merge = df_merge.drop(['name_programs', 'link_programs'], axis=1)

# drop duplicates that occur because a facility shows up both under regular and children & adolescents
df_merge = df_merge.drop_duplicates(subset=['link'])
len(df_merge)

# save dataframe
df_merge.to_csv('carf-accredited-merged.csv', index=False)
