==> facilities/carf/03-format-for-concat.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('carf-accredited-merged.csv')
df.head()

# create one massive column that we'll use to fill in our columns - fill nas with blanks so that the summed column doesn't return an NA
df['program_all'] = df.fillna('').program_offered_2 + ' / ' + df.fillna('').program_offered_3 + ' / ' + df.fillna('').program_offered_4 + ' / ' + \
       df.fillna('').program_offered_5 + ' / ' + df.fillna('').program_offered_6 + ' / ' + df.fillna('').program_offered_7 + ' / ' + \
       df.fillna('').program_offered_8 + ' / ' + df.fillna('').program_offered_9 + ' / ' + df.fillna('').program_offered_10

df = df.replace('Residential Treatment (BH) Mental Health Eating Disorders', 'Residential Treatment (BH) Mental Health Eating Disorders - Adults')

# create blank columns to fill
df[[
    # treatment types: hospital inpatient, residential, partial hospitalization day treatment, outpatient
    'hi', 'res', 'phdt', 'op',
    # facility types: 
    'cbhc', 'msmh', 'omh', 'ores', 'ph', 'psy', 'rtca', 'rtcc', 'ipsy', 'shp', 'vahc',
    # age groups: child (<12), young adult (18-25), adult (25-64), senior (65+) 
    'chld', 'yad', 'adlt', 'snr',
    # special programs offered - ped is EDs
    'tay', 'se', 'gl', 'vet', 'adm', 'mf', 'cj', 'co', 'hv', 'dv', 'trma', 'tbi', 'alz', 'ped', 'pefp', 'ptsd', 'sed', 'smi'
   ]] = None

df.columns

# treatment types
# residential
df['res'] = np.where(df.program_all.str.contains('Residential Treatment'), 1, None)
# partial hospitalization / day treatment
df['phdt'] = np.where(df.program_all.str.contains('Partial Hospitalization|Day Treatment'), 1, None)
# outpatient
df['op'] = np.where(df.program_all.str.contains('Intensive Outpatient Treatment'), 1, None)

# facility types
# msmh - Multi-setting mental health facility (e.g., non-hospital residential plus either outpatient and/or partial hospitalization/day treatment)
df['msmh'] = np.where(df.program_all.str.contains('Intensive Outpatient Treatment|Partial Hospitalization') & (df.program_all.str.contains('Residential Treatment')), 1, None)
# ohm - outpatient mental health
df['omh'] = np.where(df.program_all.str.contains('Intensive Outpatient Treatment'), 1, None)
# ph - outpatient mental health
df['ph'] = np.where(df.program_all.str.contains('Partial Hospitalization'), 1, None)

# age groups
df['chld'] = np.where(df.program_all.str.contains('Residential Treatment (BH) Mental Health Eating Disorders - Children and Adolescents'), 1, None)
df['yad'] = np.where(df.program_all.str.contains('Residential Treatment (BH) Mental Health Eating Disorders'), 1, None)
df['adlt'] = np.where(df.program_all.str.contains('Residential Treatment (BH) Mental Health Eating Disorders'), 1, None)
df['snr'] = np.where(df.program_all.str.contains('Residential Treatment (BH) Mental Health Eating Disorders'), 1, None)

# treatment types
# eating disorders
df['ped'] = np.where(df.program_all.str.contains('Eating Disorders'), 1, None)
# co-occurring mental and substance use disorders
df['co'] = np.where(df.program_all.str.contains('SUD|Substance Use Disorders|'), 1, None)

# drop the columns
df = df.drop(columns=['program_offered_1',
       'program_offered_2', 'program_offered_3', 'program_offered_4',
       'program_offered_5', 'program_offered_6', 'program_offered_7',
       'program_offered_8', 'program_offered_9', 'program_offered_10',
       'program_all'])

# split addresses into columns
df['city'] = df['address_pt2'].str.split(', ', expand=True)[0]
df['state'] = df['address_pt2'].str.split(', ', expand=True)[1]

# split state column to get zip
df['zip'] = df['state'].str.split(' ', expand=True)[2]

# replace the state column with itself, but without the zip (could also use a regex replace here)
df['state'] = df['state'].str.split(' ', expand=True)[0]

df.to_csv('carf-accredited-for-concat.csv', index=False)
